+++ CHANGELOG.md
@@ -32,6 +32,7 @@
 - [11023](https://github.com/vegaprotocol/vega/issues/11023) - Add proposed fees to `vAMM` data.
 - [11028](https://github.com/vegaprotocol/vega/issues/11028) - Add API to estimate order book depth based on `vAMM`.
 - [11400](https://github.com/vegaprotocol/vega/issues/11400) - Add support for long block auction.
+- [11026](https://github.com/vegaprotocol/vega/issues/11026) - Add API flag to get paid liquidity fees for a `vAMM` using the parent key.
 
 ### 🐛 Fixes